fix tests

+++ lib/forums/posts/deleteForumPost.ts
@@ -1,9 +1,16 @@
 import { prisma } from 'db';
 
-export async function deleteForumPost(postId: string) {
-  await prisma.post.delete({
+export async function deleteForumPost(pageId: string) {
+  const page = await prisma.page.delete({
     where: {
-      id: postId
+      id: pageId
     }
   });
+  if (page.postId) {
+    await prisma.post.delete({
+      where: {
+        id: page.postId
+      }
+    });
+  }
 }
